+++ examples/collection.py
@@ -1,6 +1,7 @@
 from helpers import authenticate
 
 from trakt import Trakt
+from trakt.objects import Movie, Show
 import logging
 import os
 
@@ -23,13 +24,23 @@ if __name__ == '__main__':
         token=authenticate()
     )
 
+    # Fetch playback for movies
+    playback = Trakt['sync/playback'].movies(exceptions=True)
+
+    for key, item in playback.items():
+        print item
+
+        if type(item) is Movie:
+            print '\tprogress: %r' % item.progress
+            print '\tpaused_at: %r' % item.paused_at
+
     # Fetch movie library (watched, collection, ratings)
     movies = {}
 
-    # Trakt['sync/watched'].movies(movies)
-    # Trakt['sync/collection'].movies(movies)
-    #
-    # Trakt['sync/ratings'].movies(movies)
+    Trakt['sync/watched'].movies(movies, exceptions=True)
+    Trakt['sync/collection'].movies(movies, exceptions=True)
+
+    Trakt['sync/ratings'].movies(movies, exceptions=True)
 
     for key, movie in movies.items():
         print movie
@@ -38,11 +49,25 @@ if __name__ == '__main__':
         print '\t', 'rating', '\t' * 3, movie.rating
 
         print '\t', 'is_watched', '\t' * 2, movie.is_watched
+        print '\t', 'last_watched_at', '\t', movie.last_watched_at
         print '\t', 'is_collected', '\t', movie.is_collected
         print '\t', 'collected_at', '\t', movie.collected_at
         print '\t', 'plays', '\t' * 3, movie.plays
 
         print
+
+    # Fetch playback for shows
+    playback = Trakt['sync/playback'].episodes(exceptions=True)
+
+    for key, item in playback.items():
+        print item
+
+        if type(item) is Show:
+            for (sk, ek), episode in item.episodes():
+                print '\t', episode
+
+                print '\t\tprogress: %r' % episode.progress
+                print '\t\tpaused_at: %r' % episode.paused_at
 
     # Fetch show/episode library (watched, collection, ratings)
     shows = {}