+++ tests/core/helpers.py
@@ -1,0 +1,22 @@
+import os
+import six
+
+if six.PY2:
+    try:
+        from six import cStringIO as BufferIO
+    except ImportError:
+        from six import StringIO as BufferIO
+else:
+    from io import BytesIO as BufferIO
+
+TESTS_PATH = os.path.abspath(os.path.dirname(__file__) + os.path.sep + '..')
+
+
+def read(path, mode='rb'):
+    if not os.path.isabs(path):
+        path = os.path.join(TESTS_PATH, path)
+
+    path = os.path.abspath(path)
+
+    with open(path, mode) as fp:
+        return BufferIO(fp.read())